Wrap TokenError with StructureException that outputs the line number.
